pyproject.toml
[project]
name = "visual-servo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyrealsense2>=2.56.5.9235",
    "ur-rtde==1.5.6",                 # 你末尾写的是 1.5.6，这里保持一致
    "numpy==1.23.5",
    "scipy==1.10.1",
    "opencv-contrib-python==4.7.0.72",
    "pyyaml==6.0.1",
    "matplotlib==3.7.1",
]